Simplify and split NManager code. Cleanup and renaming.
