+++ src/main/Anvil/Internal/ICoreSignalHandler.cs
@@ -1,0 +1,8 @@
+namespace Anvil.Internal
+{
+  public interface ICoreSignalHandler
+  {
+    void OnStart();
+    void OnShutdown();
+  }
+}
