+++ src/main/NWN/API/Async/NwTask.cs
@@ -1,8 +1,10 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.Runtime.InteropServices;
 using System.Threading;
 using System.Threading.Tasks;
+using NWN.Core;
 
 namespace NWN.API
 {
@@ -155,6 +157,18 @@ namespace NWN.API
       Task<TResult> results = await Task.WhenAny(tasks);
       await SwitchToMainThread();
       return results;
+    }
+
+    internal static async Task WaitForNativeMainThread()
+    {
+      TaskCompletionSource tcs = new TaskCompletionSource();
+      Action waitAction = () =>
+      {
+        tcs.SetResult();
+      };
+
+      VM.QueueOnMainThread(Marshal.GetFunctionPointerForDelegate(waitAction));
+      await tcs.Task;
     }
   }
 }
